+++ src/kinesis.rs
@@ -1,8 +1,8 @@
 use std::fmt::Debug;
 
+use crate::kinesis::models::*;
 use async_trait::async_trait;
 use aws_sdk_kinesis::operation::get_shard_iterator::GetShardIteratorOutput;
-use aws_sdk_kinesis::primitives::DateTime;
 use aws_sdk_kinesis::{Client, Error};
 use chrono::Utc;
 use log::{debug, error};
@@ -11,60 +11,7 @@ use tokio::sync::mpsc::Sender;
 use tokio::time::{sleep, Duration};
 
 mod helpers;
-
-#[derive(Debug, Clone)]
-pub struct ShardIteratorProgress {
-    last_sequence_id: Option<String>,
-    next_shard_iterator: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-pub enum ShardProcessorADT {
-    Termination,
-    Progress(Vec<RecordResult>),
-}
-
-#[derive(Debug, Clone)]
-pub struct PanicError {
-    pub message: String,
-}
-#[derive(Debug, Clone)]
-pub struct RecordResult {
-    pub shard_id: String,
-    pub sequence_id: String,
-    pub datetime: DateTime,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone)]
-pub struct ShardProcessorConfig {
-    pub client: Client,
-    pub stream: String,
-    pub shard_id: String,
-    pub tx_records: Sender<Result<ShardProcessorADT, PanicError>>,
-}
-
-#[derive(Debug, Clone)]
-pub struct ShardProcessorLatest {
-    pub config: ShardProcessorConfig,
-}
-
-#[derive(Debug, Clone)]
-pub struct ShardProcessorAtTimestamp {
-    pub config: ShardProcessorConfig,
-    pub from_datetime: chrono::DateTime<Utc>,
-}
-
-#[async_trait]
-pub trait ShardProcessor: Send + Sync + Debug {
-    async fn run(&self) -> Result<(), Error>;
-
-    async fn publish_records_shard(
-        &self,
-        shard_iterator: &str,
-        tx_shard_iterator_progress: Sender<ShardIteratorProgress>,
-    ) -> Result<(), Error>;
-}
+pub mod models;
 
 pub fn new(
     client: Client,
